riscv128 : updating configuration files
Setting up the configuration files in such a way that the support for
building a riscv128 cross-environment tool can be bootstrapped.
Once the patch applied the configuration can be launched smoothly:
../configure --target=riscv128-unknown-elf --enable-maintainer-mode
However, as expected, make fails and nothing works.
However, the riscv32 and riscv64 toolchains can still be built normally.