readelf.c: Add hacks to support 128-bit format
Since the removal of elf_vma (that could be typedefed approriately for
32, 64 or 128 and 128 bits), we can use neither the generic BYTE_GET
macro, nor the print_vma function.
To circumvent the issues, we add a BYTE_SETGET macro that makes two
accesses and then memcpies the field, and a PRINT_VMA macro that
handles all sizes vma printing.

Note that there are still mistakes on the output, but most segfault
are now avoided.